Added __init to api, moved jwt_ops into
